requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
bcrypt==4.1.2
python-dotenv==1.0.0
Werkzeug==3.0.1
